+++ src/ska_mid_dish_manager/devices/DishManagerDS.py
@@ -57,11 +57,11 @@ from ska_mid_dish_manager.utils.decorators import (
     record_mode_change_request,
     requires_component_manager,
 )
-from ska_mid_dish_manager.utils.schedulers import WatchdogTimerInactiveError
-from ska_mid_dish_manager.utils.track_table_input_validation import (
+from ska_mid_dish_manager.utils.input_validation import (
     TrackLoadTableFormatting,
     TrackTableTimestampError,
 )
+from ska_mid_dish_manager.utils.schedulers import WatchdogTimerInactiveError
 
 DevVarLongStringArrayType = Tuple[List[ResultCode], List[Optional[str]]]
 
@@ -136,6 +136,7 @@ class DishManager(SKAController):
             ("TrackStop", "track_stop_cmd"),
             ("ConfigureBand1", "configure_band_cmd"),
             ("ConfigureBand2", "configure_band_cmd"),
+            ("ConfigureBand", "configure_band_with_json"),
             ("Slew", "slew"),
             ("Scan", "scan"),
             ("TrackLoadStaticOff", "track_load_static_off"),
@@ -1547,6 +1548,60 @@ class DishManager(SKAController):
         return ([return_code], [message])
 
     @command(
+        dtype_in="DevString",
+        doc_in="""The command accepts a JSON string containing data to configure the SPFRx.
+        The JSON structure is as follows:
+        {
+            "receiver_band": <string>,
+            "sub_band": <integer>,
+            "spfrx_processing_parameters": {
+                "dishes": List[<string>],
+                "sync_pps":  <bool>,
+                "attenuation_pol_x": <float>,
+                "attenuation_pol_y": <float>,
+                "attenuation_1_pol_x": <float>,
+                "attenuation_1_pol_y": <float>,
+                "attenuation_2_pol_x": <float>,
+                "attenuation_2_pol_y": <float>,
+                "saturation_threshold": <float>,
+                "noise_diode": {
+                    "pseudo_random": {
+                        "binary_polynomial": <long>,
+                        "seed": <long>,
+                        "dwell": <long>,
+                    },
+                    "periodic": {
+                        "period": <long>,
+                        "duty_cycle": <long>,
+                        "phase_shift": <long>,
+                    }
+                }
+            }
+        }
+        where 'receiver_band', 'dishes' and 'sync_pps' are mandatory fields.
+        when 'receiver_band' is set to '5b', the 'sub_band' field is mandatory.
+        The dishes field is a list of dish names that the SPFRx should be configured for,
+        if 'all' is specified in the list, the SPFRx will be configured for all dishes.
+        """,
+        dtype_out="DevVarLongStringArray",
+        display_level=DispLevel.OPERATOR,
+    )
+    @BaseInfoIt(show_args=True, show_kwargs=True, show_ret=True)
+    def ConfigureBand(self, value) -> DevVarLongStringArrayType:
+        """Configure band according to JSON string supplied.
+
+        This command triggers the Dish to transition to the CONFIG Dish
+        Element Mode, and returns to the caller.
+
+        :return: A tuple containing a return code and a string
+            message indicating status.
+        """
+        handler = self.get_command_object("ConfigureBand")
+
+        result_code, unique_id = handler(value)
+        return ([result_code], [unique_id])
+
+    @command(
         dtype_in=bool,
         doc_in="If the synchronise argument is True, the SPFRx FPGA is instructed to synchronise "
         "its internal flywheel 1PPS to the SAT-1PPS for the ADC that is applicable to the band "
@@ -1561,6 +1616,8 @@ class DishManager(SKAController):
         operate in frequency band 1. On completion of the band
         configuration, Dish will automatically revert to the previous Dish
         mode (OPERATE or STANDBY‐FP).
+
+        NB. This command will be soon deprecated in favour of the generic ConfigureBand command.
 
         :return: A tuple containing a return code and a string
             message indicating status.
@@ -1587,6 +1644,8 @@ class DishManager(SKAController):
         operate in frequency band 2. On completion of the band
         configuration, Dish will automatically revert to the previous Dish
         mode (OPERATE or STANDBY‐FP).
+
+        NB. This command will be soon deprecated in favour of the generic ConfigureBand command.
 
         :return: A tuple containing a return code and a string
             message indicating status.