+++ tests/acceptance/transitions/test_ignoring_sub_devices.py
@@ -97,7 +97,7 @@ def test_ignoring_spfrx(
 
     expected_progress_updates = [
         "Fanned out commands: DS.SetIndexPosition",
-        "Awaiting configuredband, dishmode change to B2, STANDBY_FP",
+        "Awaiting configuredband change to B2",
         "ConfigureBand2 completed",
     ]